requirements.txt
openai>=1.12.0
google-api-python-client>=2.118.0
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.1
python-dotenv>=1.0.1
obsws-python>=1.6.1
websockets>=12.0
aiohttp>=3.9.3
python-multipart>=0.0.9
pydantic>=2.6.3
torch>=2.0.0
numpy>=1.21.0
transformers>=4.30.0
sounddevice>=0.4.6
style-bert-vits2>=0.1.0
pytchat>=0.5.0
# HippoRAG (長期記憶 / Retrieval-Augmented Generation)
git+https://github.com/OSU-NLP-Group/HippoRAG.git
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4
fugashi>=1.2.1
unidic-lite>=1.0.8
# 音声認識（Google Speech-to-Text）
google-cloud-speech>=2.20.0
pyaudio>=0.2.13
SpeechRecognition>=3.10.0 